node_index now returns a size_t

+++ include/dml.h
@@ -55,7 +55,7 @@ typedef uint32_t DML_SiteRank;
 typedef struct {
   /* Data distribution fixed for the entire file */
   int (*node_number)(const int coords[]);
-  int (*node_index)(const int coords[]);
+  size_t (*node_index)(const int coords[]);
   void (*get_coords)(int coords[], int node, const size_t index);
   size_t (*num_sites)(int node);
   int *latsize;
